added login and logical restrictions

+++ group/views.py
@@ -6,21 +6,28 @@ from main.models import File
 from main.upload import file_upload
 from .models import Group, Message
 from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
+
+@login_required(login_url='login')
 def show_group(request):
     user = request.user
     groups = user.group_set.all()
     context = {'groups':groups}
     return render(request, 'groups.html', context)
 
+@login_required(login_url='login')
 def show_one_group(request, pk):
     group = Group.objects.get(id=pk)
     messages = group.message_set.all()
     participants = group.participants.all()
-    context = {'group':group, 'messages':messages,'participants':participants}
+    admin = False
+    if(request.user == group.host): admin = True
+    context = {'group':group, 'messages':messages,'participants':participants, 'admin':admin}
     return render(request, 'single-group.html', context)
 
+@login_required(login_url='login')
 def create_group(request):
     if request.method == 'POST':
         host = request.user
@@ -31,13 +38,19 @@ def create_group(request):
         return redirect('groups')
     return render(request, 'create-group.html')
 
+@login_required(login_url='login')
 def delete_group(request, pk):
     group = Group.objects.get(id=pk)
+    if group.host != request.user:
+        return HttpResponse('You are not allowed to do this')
     group.delete()
     return redirect('groups')
 
+@login_required(login_url='login')
 def add_person(request, pk):
     group= Group.objects.get(id=pk)
+    if group.host != request.user:
+        return HttpResponse('You are not allowed to do this')
     if request.method == 'GET':
         return render(request, 'add-part.html', {'message':'', 'group':group})
     email = request.POST['email']
@@ -50,7 +63,7 @@ def add_person(request, pk):
 
 def username_exists(username):
     return User.objects.filter(username=username).exists()
-
+@login_required(login_url='login')
 def remove_person(request, pk, rp):
     if request.method=='GET':
         user = request.user
@@ -61,7 +74,7 @@ def remove_person(request, pk, rp):
         group.participants.remove(user2)
         return redirect(request.META.get('HTTP_REFERER'))
         
-
+@login_required(login_url='login')
 def add_file(request,pk):
     group = Group.objects.get(id=pk)
     if request.method == 'GET':
@@ -82,8 +95,10 @@ def add_file(request,pk):
     cmd = 'uploads/'+name
     os.remove(cmd)
     new.delete()
-    return redirect(request.META.get('HTTP_REFERER'))
+    path = request.META.get('HTTP_REFERER')
+    return redirect(path[0:-9])
 
+@login_required(login_url='login')
 def remove_file(request, pk, fid):
     if request.method=='GET':
         user = request.user
